% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ.R
\name{estlambda2}
\alias{estlambda2}
\title{Estimate the inflation factor for a distribution of P-values or 1df chi-square test.}
\usage{
estlambda2(p.o, p.e, plot = FALSE, filter = TRUE, adjust.xy = FALSE, ...)
}
\arguments{
\item{p.o}{Observed P-values from the data.}

\item{p.e}{Expected P-values from the NULL distribution, usually obtained through label permutations of the data.}

\item{plot}{Indicate if a plot should be produced.}

\item{filter}{Indicate if the filter should be applied. This parameter behaves the same as in estlambda in GenABEL.}

\item{adjust.xy}{Indicate if the x-axis and y-axis should be adjusted to their own range.}

\item{...}{arguments passed the plot function.}
}
\value{
Returns a list containing the estimated lambda value (estimate) and its standard deviation (se).
}
\description{
Estimate the inflation factor for a distribution of P-values or 1df chi-square test using a NULL distribution of P-values provided by user.
 The implementation is similar to the default implementation in GenABEL but does not assume the NULL distribution of P-values to be uniform.
}
\examples{
#load pre-computed p-values for igm dataset
library(QQperm)
data("example.Ps")

#print output to pdf file only if not running in interactive mode
if (!interactive()) {
 pdf("lambda.pdf")
}

#estimate inflation factor and generate plot.
lambda <-estlambda2(example.Ps$observed,example.Ps$perm, plot = TRUE, adjust.xy = TRUE)

if (!interactive()) {
 dev.off()
}
}
\author{
Slave Petrovski and Quanli Wang
}

