% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGM.R
\name{igm.read.data}
\alias{igm.read.data}
\title{Read sample file and genotype collapsing matrix in IGM (Insitute for Genomic Medicine, Columbia University) format.}
\usage{
igm.read.data(sample.file, matrix.file, filter.list = NULL)
}
\arguments{
\item{sample.file}{The input sample file.}

\item{matrix.file}{The input matrix file.}

\item{filter.list}{An optional list of genes to exclude from the analysis.}
}
\value{
Returns a list containing the data matrix (matrix) and case(1)/control(0) phenotype indicators (is.case).
}
\description{
Read sample file and genotype collapsing matrix in IGM (Insitute for Genomic Medicine, Columbia University) format.
}
\examples{
#igm.data <- igm.read.data(sample.file, matrix.file)

}
\author{
Slave Petrovski and Quanli Wang
}

