% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQperm.R
\docType{package}
\name{QQperm}
\alias{QQperm}
\alias{QQperm-package}
\alias{qqperm}
\title{QQperm: A package for permutation-based QQ plots and inflation factor estimates.}
\description{
QQperm: A package for permutation-based QQ plots and inflation factor estimates.
}
\author{
Slave Petrovski and Quanli Wang
}

