% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ.R
\name{qqplot}
\alias{qqplot}
\title{QQ plot of observed P-values vs expected P-values.}
\usage{
qqplot(P.perm, P.observed, adjust.xy = TRUE, ...)
}
\arguments{
\item{P.perm}{Expected P-values from NULL distribution, which is generated through permutation in our example.}

\item{P.observed}{Observed P-values from data.}

\item{adjust.xy}{Indicates if the x-axis and y-axis should be adjusted based on their own range in the plot.}

\item{...}{arguments passed the plot function.}
}
\value{
None.
}
\description{
QQ plot of observed P-values vs expected P-values.
}
\examples{
library(QQperm)
data("example.Ps")

#print output to pdf file only if not running in interactive mode
if (!interactive()) {
 pdf("qqplot.pdf")
}

qqplot(example.Ps$perm, example.Ps$observed)

if (!interactive()) {
 dev.off()
}
}
\author{
Slave Petrovski and Quanli Wang
}

