% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy.data}
\alias{toy.data}
\title{An example collapsing data matrix and the associated case/control status}
\format{A list containing both the genotype data matrix and the case/control statuses:

 \describe{
    \item{data}{a named data matrix, with a row for each gene and a column for each sample. Each cell in the matrix then represents an indicator function that reflects whether the given sample had none (0) or at least one (1) qualifying variant in the corresponding gene. It also accepts counts of qualifying variants for a given (gene,sample) pair, but will treat these as a (0/1) indicator.}
   \item{is.case}{case/control indicator. The order of the indicators here must match that of the order of the samples defined in the data matrix.}
}}
\usage{
toy.data
}
\description{
An example collapsing data matrix and the associated case/control status
}
\keyword{datasets}

