% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.Ps}
\alias{example.Ps}
\title{Distributions of expected and observed P-values from igm.data data set.}
\format{A list contains both expected and observed P-values.

 \describe{
    \item{perm}{A vector for expected P-values.}
   \item{observed}{A vector for observed P-values.}
}}
\source{
\url{http://ask.slave.for.it/}
}
\usage{
example.Ps
}
\description{
Distributions of expected and observed P-values from igm.data data set.
}
\keyword{datasets}

