% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGM.R
\name{igm.get.pvalues}
\alias{igm.get.pvalues}
\title{Generate NULL distribution of P-values through label switching and permutation and compute distribution of observed P-values.}
\usage{
igm.get.pvalues(matrix, is.case, n.permutations = 1000)
}
\arguments{
\item{matrix}{The input genotype matrix, with rows for genes and columns for samples.}

\item{is.case}{The case/control indicator.}

\item{n.permutations}{Number of label permutaitons.}
}
\value{
Returns A list contains observed P-values (observed) and permuted P-Values for NULL distribution (perm).
}
\description{
Generate NULL distribution of P-values through label switching and permutation and compute distribution of observed P-values.
}
\examples{
#Ps  <- igm.get.pvalues(matrix, is.case)

}
\author{
Slave Petrovski and Quanli Wang
}

