% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGM.R
\name{igm.read.data}
\alias{igm.read.data}
\title{Read sample file and genotype collipsing matrix in IGM format.}
\usage{
igm.read.data(sample.file, matrix.file, filter.list = NULL)
}
\arguments{
\item{sample.file}{The input sample file.}

\item{matrix.file}{The input matrix file.}

\item{filter.list}{A list of genes that will be excluded from the analysis.}
}
\value{
Returns A list contains data matrix (matrix) and phenotype indicators (is.case).
}
\description{
Read sample file and genotype collipsing matrix in IGM format.
}
\examples{
#igm.data <- igm.read.data(sample.file, matrix.file)

}
\author{
Slave Petrovski and Quanli Wang
}

