% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQperm.R
\docType{package}
\name{QQperm}
\alias{QQperm}
\alias{QQperm-package}
\alias{qqperm}
\title{QQperm: A package for inflation factor estimation and QQ plot.}
\description{
QQperm: A package for inflation factor estimation and QQ plot.
}
\author{
Slave Petrovski and Quanli Wang
}

