% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy.data}
\alias{toy.data}
\title{An example collapsing data matrx and associated case/control status}
\format{A list contains both genotype data matrix and case/control status:

 \describe{
    \item{data}{a named data matrix, with rows for genes and columns for samples. Each cell in the matrix represent the counts of variants for given (gene,sample) pair.}
   \item{is.case}{case/control indicator. The order of indicators must matches that of the samples defined in data matrix.}
}}
\source{
\url{http://ask.slave.for.it/}
}
\usage{
toy.data
}
\description{
An example collapsing data matrx and associated case/control status
}
\keyword{datasets}

